refactor(schemas): migrate Phase 31 endpoint schemas — batch_jobs.py 10 classes + 2 enums (#6042)
- batch_jobs.py: migrate BatchJobStatus, BatchJobType enums + BatchJobCreate,
  BatchJob, BatchTemplate, BatchSchedule, BatchJobList, LogEntry→BatchLogEntry,
  BatchRequest→APIBatchRequest, BatchResponse→APIBatchResponse → schemas_workflows.py

Renames to avoid generic-name collisions:
- LogEntry → BatchLogEntry (too generic)
- BatchRequest → APIBatchRequest (could collide with other Batch* request types)
- BatchResponse → APIBatchResponse

schemas_workflows.py: gains datetime import for batch job timestamps.
